Revert "Move navigation permissions into routing module"
This reverts commit 7e32e783d5e1e86f79d777d391ba79bcacaf643e.
