+++ apps/web/src/app/layouts/navbar.component.ts
@@ -12,7 +12,7 @@ import { Utils } from "@bitwarden/common/misc/utils";
 import { Organization } from "@bitwarden/common/models/domain/organization";
 import { Provider } from "@bitwarden/common/models/domain/provider";
 
-import { canAccessOrgAdmin } from "../organizations/organization-routing.module";
+import { canAccessOrgAdmin } from "../organizations/services/navigation-permissions.service";
 
 @Component({
   selector: "app-navbar",
